Refactor: faq 페이지에서 constants에 정의해 둔 객체를 활용하도록 변경한다
