+++ src/components/faq/QuestionList/QuestionList.component.tsx
@@ -1,12 +1,6 @@
 import { Accordion } from '@/components';
-import { Item } from '@/components/common/BulletedList/BulletedList.component';
+import { Question } from '@/constants';
 import * as Styled from './QuestionList.styled';
-
-export interface Question {
-  id: string;
-  title: string;
-  content: Item[];
-}
 
 interface QuestionListProps {
   questions: Question[];
